enable booting the guest from a local kernel directly
the user in control of a kvm-manager guest can now choose a kernel to
boot directly by placing two files (symlinks are fine) at:

  ~/vms/$VMNAME/kernel
  ~/vms/$VMNAME/initrd

This is by analogy with ~/vms/$VMNAME/cd.iso.

If those two files are present, then the kvm guest will boot directly
to the provided kernel rather than doing BIOS emulation that hands off
control to the MBR of the first emulated disks.

In this case, the first line of the contents of
/etc/sv/kvm/env/CMDLINE will be passed as the kernel parameters, and
no bootloader will be run.