update to work with jessie
around 3.15, the linux kernel stopped parsing arguments passed to it
after a -- argument:

https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=51e158c12aca3c9ac63988611a97c05109b14dc9

This breaks the usual debian-installer approach of putting kernel
arguments that you want to persist in the install after the --
argument.

We work around this by just putting the arguments we want to persist
both before and after the --.  yuck.