screen 4.5.0 and later expects an argument for -L
before stretch, screen took no argument for -L, and we relied on the
screenrc to indicate the name of the logfile.

in stretch, screen's -L takes an argument -- and it complains if the
argument following -L starts with a -.  This is a breaking API change
:/

so this patch can't go into versions before stretch.