Option to use alternative disk driver instead of default virtio
For example, to use scsi instead of virtio on the disk HDB, create an
environment variable named HDB_DRIVER and set it to scsi. You can set
this variable for any disk by changing the variable name (e.g.
HDA_DRIVER, HDC_DRIVER, etc.) to correspond to the disk for which you
want the driver set.

If you are using an SSD card and you want the fstrim command available,
set the driver to scsi. The value can be any value supported by kvm.