update attribution to be more accurate
I'm relying on the attribution lines in the README.md file
rather than trying to keep track of them in the individual files.

I may have created the individual files, but it is a lot of code
copied and pasted from other files so more accurate to show all
authors of all code in one place.